+++ AutoparkV3/AutoparkV3/Engines/AbstractEngineCE.cs
@@ -9,11 +9,11 @@ namespace AutoparkV3.Engines
 {
     public abstract class AbstractEngineCE : AbstractEngine
     {
-        public float EngineCapacity { get; set; }
-        public float FuelConsumptionPer100 { get; set; }
+        public double EngineCapacity { get; set; }
+        public double FuelConsumptionPer100 { get; set; }
         public int FuelTankCapacity { get; set; }
 
-        public override float GetMaxKilometers()
+        public override double GetMaxKilometers()
         {
             return FuelTankCapacity * 100 / FuelConsumptionPer100;
         }
