+++ AutoparkV3/AutoparkV3/Engines/EngineGasoline.cs
@@ -13,7 +13,7 @@ namespace AutoparkV3.Engines
             Init();
         }
 
-        public EngineGasoline(float engineCapacity, float fuelConsumptionPer100, int fuelTankCapacity)
+        public EngineGasoline(double engineCapacity, double fuelConsumptionPer100, int fuelTankCapacity)
         {
             Init();
             EngineCapacity = engineCapacity;
